data_update_requirements.txt
requests
pandas
mysql-connector-python
sqlalchemy
python-dotenv
